requirements.txt
torch>=1.9.0
torchvision>=0.10.0
segmentation-models-pytorch>=0.3.0
Pillow>=8.0.0
numpy>=1.21.0
albumentations>=1.3.0
timm>=0.9.0
pandas>=1.3.0
transformers>=4.20.0